backend/requirements.txt
google-genai>=1.0.0
pyaudio>=0.2.14
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
